# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/comment/update/{commentId}": {
      "put": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "updateComment",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "newContent",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/register": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/followrelations/unfollow": {
      "post": {
        "tags": [
          "follow-relations-controller"
        ],
        "operationId": "unFollow",
        "parameters": [
          {
            "name": "whoIsFollowing",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "whoIsBeingFollowed",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/followrelations/follow": {
      "post": {
        "tags": [
          "follow-relations-controller"
        ],
        "operationId": "follow",
        "parameters": [
          {
            "name": "whoIsFollowing",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "whoIsBeingFollowed",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/comment/create": {
      "post": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "createComment",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "articleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "content",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentCommentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": -1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/{userId}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/likes": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getUserLikes",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/followrelations/followingto": {
      "get": {
        "tags": [
          "follow-relations-controller"
        ],
        "operationId": "findFollowing",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/followrelations/followersof": {
      "get": {
        "tags": [
          "follow-relations-controller"
        ],
        "operationId": "findFollowers",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comment/{commentId}": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "findByCommentId",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDTO"
                }
              }
            }
          }
        }
      }
    },
    "/comment/user/{userId}": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "getCommentsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comment/child": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "getAllChildComments",
        "parameters": [
          {
            "name": "parentCommentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comment/article/{articleId}": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "getCommentsByArticleId",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bookmark/user/{userId}": {
      "get": {
        "tags": [
          "bookmarked-controller"
        ],
        "operationId": "getBookmarkedArticlesByUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/article/{articleId}": {
      "get": {
        "tags": [
          "article-controller"
        ],
        "operationId": "getArticleById",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/article/comments/{articleId}": {
      "get": {
        "tags": [
          "article-controller"
        ],
        "operationId": "getCommentsForArticle",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/comment/delete/{commentId}": {
      "delete": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "deleteComment",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArticleDTO": {
        "type": "object",
        "properties": {
          "articleId": {
            "type": "integer",
            "format": "int32"
          },
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "publishedDate": {
            "type": "string",
            "format": "date-time"
          },
          "premiumRequired": {
            "type": "boolean"
          },
          "content": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "tags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDTO"
            }
          },
          "likes": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LikesDTO"
            }
          },
          "bookmarks": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookmarkedDTO"
            }
          },
          "comments": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDTO"
            }
          }
        }
      },
      "BookmarkedDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userDTO": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "articleDTO": {
            "$ref": "#/components/schemas/ArticleDTO"
          }
        }
      },
      "CommentDTO": {
        "type": "object",
        "properties": {
          "commentId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "userDTO": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "articleDTO": {
            "$ref": "#/components/schemas/ArticleDTO"
          },
          "parentCommentId": {
            "type": "integer",
            "format": "int64"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FollowRelationsDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "whoIsFollowing": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "whoIsBeingFollowed": {
            "$ref": "#/components/schemas/UserDTO"
          }
        }
      },
      "LikesDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userDTO": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "articleDTO": {
            "$ref": "#/components/schemas/ArticleDTO"
          }
        }
      },
      "TagDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tag": {
            "type": "string"
          },
          "articleDTO": {
            "$ref": "#/components/schemas/ArticleDTO"
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "emailId": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "likes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LikesDTO"
            }
          },
          "bookmarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookmarkedDTO"
            }
          },
          "followers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FollowRelationsDTO"
            }
          },
          "following": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FollowRelationsDTO"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDTO"
            }
          },
          "subscribed": {
            "type": "boolean"
          },
          "bio": {
            "type": "string"
          }
        }
      }
    }
  }
}